增加全文搜索服务

+++ shared/Zhaoxi.MSACommerce.SharedKernel/Paging/PagedList.cs
@@ -2,7 +2,7 @@
 
 public class PagedList<T> : List<T>
 {
-    public PagedList(IEnumerable<T> items, int count, Pagination pagination)
+    public PagedList(IEnumerable<T> items, long count, Pagination pagination)
     {
         MetaData = new PagedMetaData
         {
